Downloads/innerve-x-main/innerve-x-main/requirements.txt
# DhanSathi Backend Dependencies
# ================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication
pyjwt==2.8.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0

# Validation
pydantic[email]==2.5.2

# AI Providers (optional - install as needed)
google-generativeai>=0.8.5  # For Gemini
# openai==1.3.7               # For OpenAI

# OCR (optional)
# pytesseract==0.3.10
# pillow==10.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
